small corrections

+++ include/rts/UnitType.h
@@ -11,7 +11,7 @@ namespace rts {
   class UnitType {
   public:
     std::array<abilities::Instance, MaxUnitAbilities> abilities{};
-    std::array<AbilityId, size_t(RelativeContent::Count)> defaultAbility{};
+    std::array<AbilityId, uint32_t(RelativeContent::Count)> defaultAbility{};
     ResourceQuantityList cost;
     ResourceQuantityList provision;
     GameTime buildTime;
