+++ include/rts/types.h
@@ -98,8 +98,8 @@ namespace rts {
   using util::geo::Rectangle;
   using util::geo::Vector;
 
-  enum class Direction : size_t { Left, Right, Up, Down };
-  enum class RelativeContent : size_t { Friend, Foe, Ground, Resource, Count };
+  enum class Direction : uint32_t { Left, Right, Up, Down };
+  enum class RelativeContent : uint32_t { Friend, Foe, Ground, Resource, Count };
 
   class Ui {
   public:
